refactore: 인프라 패키지 구조 변경
