+++ src/main/java/com/swyp/index/IndexApplication.java
@@ -11,5 +11,4 @@ public class IndexApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(IndexApplication.class, args);
 	}
-
 }
