+++ src/main/java/com/swyp/index/application/user/AuthService.java
@@ -14,7 +14,7 @@ import com.swyp.index.domain.user.User;
 import com.swyp.index.global.exception.CustomException;
 import com.swyp.index.global.exception.ErrorCode;
 import com.swyp.index.global.security.JwtProvider;
-import com.swyp.index.infrastructure.user.UserRepository;
+import com.swyp.index.infrastructure.repository.UserRepository;
 import com.swyp.index.presentation.dto.user.LoginRequest;
 import com.swyp.index.presentation.dto.user.SignUpRequest;
