+++ src/main/java/com/swyp/index/application/user/CustomOAuth2UserService.java
@@ -19,7 +19,7 @@ import com.swyp.index.domain.user.NaverUserInfo;
 import com.swyp.index.domain.user.OAuth2UserInfo;
 import com.swyp.index.domain.user.Provider;
 import com.swyp.index.domain.user.User;
-import com.swyp.index.infrastructure.user.UserRepository;
+import com.swyp.index.infrastructure.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
