+++ src/main/java/com/swyp/index/application/user/CustomUserDetailsService.java
@@ -11,7 +11,7 @@ import com.swyp.index.domain.user.Provider;
 import com.swyp.index.domain.user.User;
 import com.swyp.index.global.exception.CustomException;
 import com.swyp.index.global.exception.ErrorCode;
-import com.swyp.index.infrastructure.user.UserRepository;
+import com.swyp.index.infrastructure.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
