+++ src/main/java/com/swyp/index/application/user/TokenService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.swyp.index.domain.user.User;
 import com.swyp.index.global.exception.CustomException;
 import com.swyp.index.global.exception.ErrorCode;
-import com.swyp.index.infrastructure.user.UserRepository;
+import com.swyp.index.infrastructure.repository.UserRepository;
 import com.swyp.index.global.security.JwtProvider;
 
 import lombok.RequiredArgsConstructor;
